Fix category icons and remove category links
Currently we only show a warning/error icon for the Performance
category. We should show an icon for all of the categories.

Additionally, the category links to the reports are currently broken,
because of a missing /reports/ in the URL. Even if we add that, though,
unfortunately these anchor links won't work as intended because the page
content is loaded dynamically by JS, which doesn't scroll to the URL
anchor. For that reason I've removed these links here.

To test: yarn && yarn serve