Better handling of URLs without trailing slash
The current logic assumes that all URLs will end in a slash and thus
generate report JSON files that look like URL_-TIMESTAMP.report.json.
URLs that don't have a slash don't include that underscore, which breaks
the create-report-index script (as reported on issue 22).

This commit modifies the report core so that a URL's trailing slash is
included in its report URL. This will impact live URLs, meaning that
instead of pages like

https://cfpb.github.io/cfgov-lighthouse/www_consumerfinance_gov-_about_us_blog/

we'll have URLs like

https://cfpb.github.io/cfgov-lighthouse/www_consumerfinance_gov-_about_us_blog_/

(note the added trailing slash).

This is probably more correct to account for hypothetical cases where
you have two URLs that differ depending on the trailing slash.