Better handling of tested URLs with query strings
As reported in issue 22, the report generation scripts currently don't
work properly if tested URLs have query strings. The logic we use to
toggle between desktop and mobile testing involves adding a ?mobile=1 to
the tested URL.

The code that runs the test handles this properly if the test URL
already has a query string:

https://github.com/cfpb/cfgov-lighthouse/blob/426ae5c6e1f040e02d08b3c53b22de10b4ce885a/scripts/lib/urls.js#L38-L44

but the code that collects the reports for the web viewer currently does
not:

https://github.com/cfpb/cfgov-lighthouse/blob/426ae5c6e1f040e02d08b3c53b22de10b4ce885a/scripts/lib/reports.js#L89-L91

This logic will fail if the URL has a different query string that
doesn't exactly start with ?mobile=1.

This commit fixes that logic by using the URL Web API interface to
modify/check the URL being tested.