Properly test using desktop preset
In the most recent versions of Lighthouse, the configuration settings
we were using for desktop testing no longer work. Instead of using
formFactor and screenEmulation, we should use the "desktop" preset.

This commit makes that change which properly tests using a desktop form
factor; this can be confirmed by looking at the (small) screenshots
included as part of a Lighthouse report.

With this change, our desktop Performance scores now are in the 90s,
which is more consistent with how things work when you run from Chrome
developer tools.