Load test URLs dynamically from file
This commit modifies which URLs are tested each night. Instead of having
a hardcoded list in urls.js, the code now tries to load the list from a
text file named urls.txt.

The nightly GitHub Action now downloads this file from a URL which needs
to be configured as a GitHub repository secret named TEST_PAGES_URL.

If the file does not exist, running `yarn lighthouse` will now only run
against the root page of the domain.