Limit report history to last 30 days
This change modifies the existing report processing script so that it
only keeps reports from the last 30 days. This prevents the size of
this repository from growing in an unbounded manner.

Currently, the repo size exceeds 12GB. For each individual page test,
Lighthouse generates an X00 KB JSON file. We have ~19,000 of these
(roughly 25 pages x 365 days x 2 reports for mobile/desktop) at this
point. The repository also contains the rendered HTML pages for each
report.

GitHub's recommended size limit for repositories is discussed here:
https://stackoverflow.com/a/59479166

Because the repository is so large, the nightly test-and-commit jobs
have been failing, resulting in the dashboard being out of date since
December.

Failing jobs:
https://github.com/cfpb/cfgov-lighthouse/actions/workflows/pages/pages-build-deployment
Dashboard link:
https://cfpb.github.io/cfgov-lighthouse/

With this change, limiting report history to the last 30 days, the
repository is reduced to < 5 GB.