Drop old URLs from dashboard report
This commit modifies how Webpack is used to generate the Lighthouse
dashboard, so that it drops old URLs that were tested in previous
nightly runs but weren't tested in the latest run.

This doesn't remove those old reports from the repository, or from the
singular reports.json file that we generate that aggregates all of the
reports -- it only modifies how the dashboard website is built to
exclude those results from the dashboard.

See internal https://github.local/Design-Development/Design-and-Content-Team/issues/502
for context.