Add querystring param to cache-bust pages URL
The nightly Lighthouse run pulls down a list of URLs to test from a
URL that is configured via a GitHub Secret. This secret value currently
points to a GitHub gist raw URL, e.g.

https://gist.githubusercontent.com/username/abcdef123456/raw/

Unfortunately, this URL doesn't always return the latest version of the
gist, apparently due to some caching on GitHub's side. To try to
bypass this caching, we can try to add a random querystring to the URLs
being fetched, e.g. ?12345678.